Find the event id for the desired event.
DCI requires two query parameters to view the scores from a specific
performance. `event`, which is some arbitrary alphanumeric value, and `year`,
which is just a string for the current year.

The `year` value can either be computed or stored as a variable, but `event`
can only be found by parsing HTML from `dci.org/scores` and matching part of
the event's name and location to the corresponding `<option>` element's text.
Once you find that element, extracting its `value` attribute will give you the
event value needed to request the correct score page.